--- built/v2-fetch-tcg.js
@@ -1,48 +1,0 @@
-import * as constants from './v2-constants.js';
-import * as get from './v2-get.js';
-import * as store from './v2-store.js';
-import * as types from './v2-types.js';
-import * as utils from './v2-utils.js';
-export async function fetchAndStoreCardsBySet(setId) {
-    const data = await fetchJson(`${constants.CARDS_SETID_URL}${setId}`);
-    const cards = data.map((row) => ({
-        id: row.id,
-        name: row.name,
-        set: row.set.id,
-        subtype: get.getSubtype(row),
-        energy: get.getEnergyType(row),
-        rarity: row.rarity,
-        filename: row.images.large,
-        supertype: row.supertype,
-        artist: row.artist,
-        nationalDex: row.nationalPokedexNumbers,
-    }));
-    store.storeCardsBySetId(setId, cards);
-    return cards;
-}
-export async function fetchAndStoreSetMetadata() {
-    const data = await fetchJson('https://api.pokemontcg.io/v2/sets');
-    store.storeSetMetaData(data);
-}
-export async function fetchJson(url) {
-    const apiKey = get.getTcgApiKey();
-    try {
-        utils.toggleStatusModal(`Fetching ${url}`, 'showstatus');
-        const response = await fetch(url, {
-            method: 'GET',
-            headers: {
-                'X-Api-Key': apiKey,
-            },
-        });
-        if (!response.ok) {
-            throw new Error(`Failed to fetch data from ${url}`);
-        }
-        const data = await response.json();
-        utils.toggleStatusModal('', 'hide');
-        return data.data;
-    }
-    catch (error) {
-        console.error(error);
-    }
-}
-//# sourceMappingURL=v2-fetch-tcg.js.map
