+++ built/v2-sort.js
@@ -3,7 +3,7 @@ import * as get from './v2-get.js';
 import * as pull from './v2-pull-fn.js';
 import * as sort from './v2-sort.js';
 import * as store from './v2-store.js';
-import * as tcg from './v2-fetch-tcg.js';
+import * as tcg from './v2-api-tcg.js';
 import * as types from './v2-types.js';
 import * as ui from './v2-ui.js';
 import * as utils from './v2-utils.js';
