+++ v2-pull.html
@@ -59,9 +59,9 @@
       </div>
     </main>
 
-    <script defer type="text/javascript" src="v2-secrets.js"></script>
-    <script defer type="text/javascript" src="v2-gapi-init.js"></script>
-    <script defer type="text/javascript" src="v2-gapi-fetch.js"></script>
+    <script defer type="text/javascript" src="v2-api-gapi-secrets.js"></script>
+    <script defer type="text/javascript" src="v2-api-gapi-init.js"></script>
+    <script defer type="text/javascript" src="v2-api-gapi-sheets.js"></script>
     <script
       defer
       src="https://apis.google.com/js/api.js"
